dusk: float internalResolutionScale, seeded before aurora init
Make game.internalResolutionScale a float so low-end handhelds can
render below the VI resolution (0.5 = 320x240 on this 640x480 game,
upscaled at present); whole values remain the integer upscale factors
and the graphics-tuner UI still steps whole multipliers.

Seed VISetFrameBufferScale from the setting before aurora_initialize so
render targets are created at the configured size from the first frame.
Applying it only post-init forced a mid-session resize, which recreates
render targets from the game thread — on Mali/GLES that takes Dawn's EGL
context while the render worker may hold it.

+++ include/dusk/settings.h
@@ -203,7 +203,10 @@ struct UserSettings {
         ConfigVar<bool> disableWaterRefraction;
         ConfigVar<bool> enableTextureReplacements;
         ConfigVar<FrameInterpMode> enableFrameInterpolation;
-        ConfigVar<int> internalResolutionScale;
+        // Float so low-end handhelds can render below the VI resolution
+        // (0.5 = 320x240 on a 640x480 VI mode); whole values remain the
+        // integer upscale factors, 0 = window-native.
+        ConfigVar<float> internalResolutionScale;
         ConfigVar<int> shadowResolutionMultiplier;
         ConfigVar<Resampler> resampler;
         ConfigVar<bool> enableMapBackground;
